scripts/requirements.txt
# Python requirements for Color Spectrum Analysis Service
# Install with: pip install -r requirements.txt

# Core dependencies
Pillow>=10.0.0
numpy>=1.24.0

# Optional but recommended for better color clustering
scikit-learn>=1.3.0
opencv-python>=4.8.0

# For API integration (optional)
# flask>=2.3.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
